remove inaccurate docs
Summary:
Our docs suggest using this babel plugin to remove console.log statements in production. Unfortunately, it does not actually work when you run it on a new React Native project, and the root cause is that the plugin does not handle all cases correctly. See discussion in https://github.com/facebook/react-native/issues/10412 . For now, we should just stop recommending that people use this plugin, because it doesn't work.
Closes https://github.com/facebook/react-native/pull/12315

Differential Revision: D4538567

Pulled By: hramos

fbshipit-source-id: f3ae1b9143130a05601907ee902a02fc0b2818b0