Remove Appetize embedded examples
Summary:
These examples work sporadically, and are all based on a years-old build of UIExplorer that was once uploaded to Appetize. As a result, newer APIs such as FlatList are not supported and attempting to run their associated examples will result in an error.

<img width="526" alt="screen shot 2017-04-24 at 12 21 31 pm" src="https://cloud.githubusercontent.com/assets/165856/25348155/57016f4a-28eb-11e7-92b0-a918ef1a0e5c.png">

This PR removes the prompts to run the examples.

See #13644 for exploration into using `<SnackPlayer>` for these.

Fixes #12300
Closes https://github.com/facebook/react-native/pull/13644

Differential Revision: D4939653

Pulled By: hramos

fbshipit-source-id: 11cf564d2ceb1981d3ce700aaae40dc873b183ad